requirements.txt
altgraph==0.17
appdirs==1.4.3
asgiref==3.2.10
auto-py-to-exe==2.7.8
bottle==0.12.18
bottle-websocket==0.2.9
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.9.2
django-heroku==0.3.1
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mysql-connector-python==8.0.18
numpy==1.19.1
opencv-python==4.3.0.36
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
protobuf==3.10.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyparsing==2.4.6
pytoml==0.1.21
python-decouple
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.13.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==16.7.7
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.1.2

